Added and fixed the Thumb LDRB_1, LDR_1, STRB_1, STR_1, LDRH_1 and STRH_1 opcodes in the interpreter.
We hit a big milestone today: The test ROM now actually loads user code and runs without showing any obvious errors.
Next Step: GPU and Video Output support

+++ src/Interpreter/ThumbInterpreters.cpp
@@ -537,13 +537,13 @@ void Interpreter::HandleThumbLoadStoreImmediateOffsetInstruction(std::shared_ptr
             _cpu->GetMemory()->WriteUInt8(Rn + instr->GetImmediate(), Rd);
             return; // Don't update the Rd register
         case Thumb::ThumbOpcodes::STR_1:
-            _cpu->GetMemory()->WriteUInt8(Rn + instr->GetImmediate(), Rd);
+            _cpu->GetMemory()->WriteUInt32(Rn + instr->GetImmediate(), Rd);
             return; // Don't update the Rd register
         case Thumb::ThumbOpcodes::LDRH_1:
-            Rd = _cpu->GetMemory()->ReadUInt32(Rn + instr->GetImmediate());
+            Rd = _cpu->GetMemory()->ReadUInt16(Rn + instr->GetImmediate());
             break;
         case Thumb::ThumbOpcodes::STRH_1:
-            _cpu->GetMemory()->WriteUInt32(Rn + instr->GetImmediate(), Rd);
+            _cpu->GetMemory()->WriteUInt16(Rn + instr->GetImmediate(), Rd);
             return; // Don't update the Rd register
         default:
             break;
